upon new connection of client on proxy_server, it will save the client info to database, as well as remove the client upon disconnection

+++ migrations/2020-01-07-125133_accounts/up.sql
@@ -4,6 +4,6 @@
 CREATE TABLE Clients(
     uid BIGINT NOT NULL CHECK (uid > 0),
     ip varchar(15) NOT NULL,
-    port SMALLINT NOT NULL,
+    port INTEGER NOT NULL,
     PRIMARY KEY (uid)
 )
